# backend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nolus Protocol Webapp API",
    "description": "Public REST API backing app.nolus.io. Read endpoints are cached and free to call; write endpoints return unsigned transaction payloads for client-side signing. Admin endpoints require API key auth and are intentionally omitted from this spec.",
    "contact": {
      "name": "Nolus Protocol",
      "url": "https://nolus.io/"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://app.nolus.io",
      "description": "Production"
    },
    {
      "url": "https://app-dev.nolus.io",
      "description": "Staging"
    }
  ],
  "paths": {
...
    "/api/earn/withdraw": {
      "post": {
        "tags": [
          "earn"
        ],
        "summary": "Build a withdraw transaction",
        "description": "Returns unsigned CosmWasm execute messages to burn nLPN receipt tokens and\nwithdraw the underlying LPN value from an earn pool.",
        "operationId": "withdraw",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unsigned withdraw transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarnTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Cache not yet populated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/batch/loans-stats": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "Loans stats batch",
        "description": "Aggregates loan stats endpoints into one response, served from\nbackground-refreshed cache. Each field is an opaque ETL passthrough.",
        "operationId": "batch_loans_stats",
        "responses": {
          "200": {
            "description": "Batch of opaque ETL passthroughs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoansStatsBatch"
                }
              }
            }
          },
          "503": {
            "description": "Cache not yet populated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/batch/stats-overview": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "Stats overview batch",
        "description": "Aggregates several ETL stats endpoints into one response, served from\nbackground-refreshed cache. Each field is an opaque ETL passthrough.",
        "operationId": "batch_stats_overview",
        "responses": {
          "200": {
            "description": "Batch of opaque ETL passthroughs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOverviewBatch"
                }
              }
            }
          },
          "503": {
            "description": "Cache not yet populated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/batch/user-dashboard": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "User dashboard batch",
        "description": "Aggregates earnings, realized PnL, and position debt for an address in\nparallel. Each field is an opaque ETL passthrough.",
        "operationId": "batch_user_dashboard",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "Nolus bech32 address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Batch of opaque ETL passthroughs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDashboardBatch"
                }
              }
            }
          },
          "502": {
            "description": "Upstream ETL call failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/batch/user-history": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "User history batch",
        "description": "Aggregates history stats and realized PnL data for an address in parallel.\nEach field is an opaque ETL passthrough.",
        "operationId": "batch_user_history",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "Nolus bech32 address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Batch of opaque ETL passthroughs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserHistoryBatch"
                }
              }
            }
          },
          "502": {
            "description": "Upstream ETL call failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/subscribe": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Subscribe to push notifications",
        "description": "Opaque passthrough to the ETL `/subscribe` endpoint. Request and response\nbodies are opaque JSON — shapes are not fixed in this spec.",
        "operationId": "proxy_subscribe",
        "requestBody": {
          "description": "Opaque ETL subscribe payload",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Opaque ETL API passthrough",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "502": {
            "description": "Upstream ETL call failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/etl/{path}": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "Generic ETL passthrough",
        "description": "Forwards GET requests to the upstream ETL API for a fixed allowlist of\npaths. Query parameters are passed through verbatim. Response is an opaque\nETL API passthrough — shape is not fixed in this spec.",
        "operationId": "etl_proxy_generic",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "Target ETL endpoint (allowlisted server-side)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opaque ETL API passthrough",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "ETL endpoint not in allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream ETL call failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EarnTransactionResponse": {
        "type": "object",
        "required": [
          "messages",
          "memo"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Unsigned CosmWasm execute messages for client-side signing"
          },
          "memo": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `VALIDATION_FAILED`, `NOT_FOUND`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message."
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "Field name that failed validation (when applicable)."
          },
          "details": {
            "type": "object",
            "description": "Additional structured details about the error."
          },
          "retry_after": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Number of seconds the client should wait before retrying (rate-limit responses).",
            "minimum": 0
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error response body",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "LoansStatsBatch": {
        "type": "object",
        "description": "Batch response for loans stats (raw JSON passthrough)",
        "required": [
          "open_position_value",
          "open_interest"
        ],
        "properties": {
          "open_position_value": {
            "type": [
              "object",
              "null"
            ]
          },
          "open_interest": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
...
      "StatsOverviewBatch": {
        "type": "object",
        "description": "Batch response for stats overview page (raw JSON passthrough)",
        "required": [
          "tvl",
          "tx_volume",
          "buyback_total",
          "realized_pnl_stats",
          "revenue"
        ],
        "properties": {
          "tvl": {
            "type": [
              "object",
              "null"
            ]
          },
          "tx_volume": {
            "type": [
              "object",
              "null"
            ]
          },
          "buyback_total": {
            "type": [
              "object",
              "null"
            ]
          },
          "realized_pnl_stats": {
            "type": [
              "object",
              "null"
            ]
          },
          "revenue": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
...
      "UserDashboardBatch": {
        "type": "object",
        "description": "Batch response for user dashboard data (raw JSON passthrough)",
        "required": [
          "earnings",
          "realized_pnl",
          "position_debt_value"
        ],
        "properties": {
          "earnings": {
            "type": [
              "object",
              "null"
            ]
          },
          "realized_pnl": {
            "type": [
              "object",
              "null"
            ]
          },
          "position_debt_value": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
      "UserHistoryBatch": {
        "type": "object",
        "description": "Batch response for user history page (raw JSON passthrough)",
        "required": [
          "history_stats",
          "realized_pnl_data"
        ],
        "properties": {
          "history_stats": {
            "type": [
              "object",
              "null"
            ]
          },
          "realized_pnl_data": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
...
      "WithdrawRequest": {
        "type": "object",
        "required": [
          "protocol",
          "amount"
        ],
        "properties": {
          "protocol": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "description": "Amount in nLPN to withdraw (receipt tokens)"
          }
        }
      },
...
    }
  },
...
}